refactor(commands): split commands.rs into feature submodules
